+++ lib/graphiti.rb
@@ -124,6 +124,8 @@ require "graphiti/sideload/polymorphic_belongs_to"
 require "graphiti/resource"
 require "graphiti/resource_proxy"
 require "graphiti/request_validator"
+require "graphiti/request_validators/validator"
+require "graphiti/request_validators/update_validator"
 require "graphiti/query"
 require "graphiti/scope"
 require "graphiti/deserializer"
